# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Auth Service.",
        "title": "Auth Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
...
        "/api/reservations": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all reservations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "List Reservations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation_service.ListReservationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "Create Reservation",
                "parameters": [
                    {
                        "description": "Create Reservation",
                        "name": "Reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation_service.CreateReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation_service.CreateReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/reservations/check": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Check the status of a reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "Check Reservation",
                "parameters": [
                    {
                        "description": "Check Reservation",
                        "name": "Check",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation_service.CheckReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation_service.CheckReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/reservations/order": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Order meals for a reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "Order Meals",
                "parameters": [
                    {
                        "description": "Order Meals",
                        "name": "Order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation_service.OrderMealsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation_service.OrderMealsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/reservations/{reservation-id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a specific reservation by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "Get Reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservation-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation_service.GetReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "Update Reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservation-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Reservation",
                        "name": "Reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation_service.UpdateReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation_service.UpdateReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a specific reservation by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "Delete Reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservation-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation_service.DeleteReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "reservation_service.CheckReservationRequest": {
            "type": "object",
            "properties": {
                "reservation_time": {
                    "type": "string"
                },
                "restaurant_id": {
                    "type": "string"
                }
            }
        },
        "reservation_service.CheckReservationResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                }
            }
        },
...
        "reservation_service.CreateReservationRequest": {
            "type": "object",
            "properties": {
                "reservation_time": {
                    "type": "string"
                },
                "restaurant_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "reservation_service.CreateReservationResponse": {
            "type": "object",
            "properties": {
                "reservation": {
                    "$ref": "#/definitions/reservation_service.Reservation"
                }
            }
        },
...
        "reservation_service.DeleteReservationResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "reservation_service.GetReservationResponse": {
            "type": "object",
            "properties": {
                "reservation": {
                    "$ref": "#/definitions/reservation_service.Reservation"
                }
            }
        },
...
        "reservation_service.ListReservationsResponse": {
            "type": "object",
            "properties": {
                "reservations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reservation_service.Reservation"
                    }
                }
            }
        },
        "reservation_service.MealOrder": {
            "type": "object",
            "properties": {
                "menu_item_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "reservation_service.OrderMealsRequest": {
            "type": "object",
            "properties": {
                "meals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reservation_service.MealOrder"
                    }
                },
                "reservation_id": {
                    "type": "string"
                }
            }
        },
        "reservation_service.OrderMealsResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "reservation_service.Reservation": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "reservation_time": {
                    "type": "string"
                },
                "restaurant_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "reservation_service.UpdateReservationRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "reservation_time": {
                    "type": "string"
                },
                "restaurant_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "reservation_service.UpdateReservationResponse": {
            "type": "object",
            "properties": {
                "reservation": {
                    "$ref": "#/definitions/reservation_service.Reservation"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}